NXP-8641: allow JSF UI to list files > 2GB

+++ nuxeo-platform-ui-web/src/main/java/org/nuxeo/ecm/platform/ui/web/tag/fn/Functions.java
@@ -336,8 +336,8 @@ public final class Functions {
 
     public static String printFormatedFileSize(String sizeS, String format,
             Boolean isShort) {
-        Integer size = (sizeS == null || "".equals(sizeS)) ? 0
-                : Integer.parseInt(sizeS);
+        long size = (sizeS == null || "".equals(sizeS)) ? 0
+                : Long.parseLong(sizeS);
         BytePrefix prefix = Enum.valueOf(BytePrefix.class, format);
         int base = prefix.getBase();
         String[] suffix = isShort ? prefix.getShortSuffixes()
